+++ RSDKv5/RSDK/Core/Legacy/v3/RetroEnginev3.cpp
@@ -784,7 +784,7 @@ int32 RSDK::Legacy::v3::LoadXMLStages(int32 mode, int32 gcStageCount)
 
                         SceneListInfo *list = &sceneInfo.listCategory[l];
                         if (!mode) {
-                            sprintf_s(list->name, (int32)sizeof(list->name), "%s", categoryNames[l]);
+                            sprintf_s(list->name, sizeof(list->name), "%s", categoryNames[l]);
                             GEN_HASH_MD5(list->name, list->hash);
 
                             list->sceneOffsetStart = gcStageCount;
@@ -812,10 +812,10 @@ int32 RSDK::Legacy::v3::LoadXMLStages(int32 mode, int32 gcStageCount)
 
                                     SceneListEntry *scene = &sceneInfo.listData[gcStageCount];
 
-                                    sprintf_s(scene->name, (int32)sizeof(scene->name), "%s", stgName);
+                                    sprintf_s(scene->name, sizeof(scene->name), "%s", stgName);
                                     GEN_HASH_MD5(scene->name, scene->hash);
-                                    sprintf_s(scene->folder, (int32)sizeof(scene->folder), "%s", stgFolder);
-                                    sprintf_s(scene->id, (int32)sizeof(scene->id), "%s", stgID);
+                                    sprintf_s(scene->folder, sizeof(scene->folder), "%s", stgFolder);
+                                    sprintf_s(scene->id, sizeof(scene->id), "%s", stgID);
 
                                     scene->filter = 0xFF;
