+++ RSDKv5/RSDK/Graphics/DX11/DX11RenderDevice.cpp
@@ -873,7 +873,7 @@ void RenderDevice::LoadShader(const char *fileName, bool32 linear)
 
     ShaderEntry *shader = &shaderList[shaderCount];
     shader->linear      = linear;
-    sprintf_s(shader->name, (int32)sizeof(shader->name), "%s", fileName);
+    sprintf_s(shader->name, sizeof(shader->name), "%s", fileName);
 
     const D3D_SHADER_MACRO defines[] = {
 #if RETRO_REV02
@@ -889,7 +889,7 @@ void RenderDevice::LoadShader(const char *fileName, bool32 linear)
     size_t bytecodeSize = 0;
 
     // Try to compile the vertex shader source if it exists
-    sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Shaders/DX11/%s.hlsl", fileName);
+    sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Shaders/DX11/%s.hlsl", fileName);
     InitFileInfo(&info);
     if (LoadFile(&info, fullFilePath, FMODE_RB)) {
         uint8 *fileData = NULL;
@@ -946,7 +946,7 @@ void RenderDevice::LoadShader(const char *fileName, bool32 linear)
     else {
 #endif
         // if the vertex shader source doesn't exist, fall back and try to load the vertex shader bytecode
-        sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Shaders/CSO-DX11/%s.vso", fileName);
+        sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Shaders/CSO-DX11/%s.vso", fileName);
         InitFileInfo(&info);
         if (LoadFile(&info, fullFilePath, FMODE_RB)) {
             uint8 *fileData = NULL;
@@ -1008,7 +1008,7 @@ void RenderDevice::LoadShader(const char *fileName, bool32 linear)
 
 #if !RETRO_USE_ORIGINAL_CODE
     // Try to compile the pixel shader source if it exists
-    sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Shaders/DX11/%s.hlsl", fileName);
+    sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Shaders/DX11/%s.hlsl", fileName);
     InitFileInfo(&info);
     if (LoadFile(&info, fullFilePath, FMODE_RB)) {
         uint8 *fileData = NULL;
@@ -1061,7 +1061,7 @@ void RenderDevice::LoadShader(const char *fileName, bool32 linear)
     else {
 #endif
         // if the pixel shader source doesn't exist, fall back and try to load the pixel shader bytecode
-        sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Shaders/CSO-DX11/%s.fso", fileName);
+        sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Shaders/CSO-DX11/%s.fso", fileName);
         InitFileInfo(&info);
         if (LoadFile(&info, fullFilePath, FMODE_RB)) {
             uint8 *fileData = NULL;
